fix(types): integrate rich sender into default sender interface
